# underline, strikethrough, insertion sort for hits is faster, context.globalAlpha