--- a/src/pixie/demo.nim
+++ b/src/pixie/demo.nim
@@ -1,6 +1,10 @@
-import opengl, pixie, pixie/context, staticglfw except Image
+import opengl, pixie, pixie/context
 
-export pixie, staticglfw except Image
+import staticglfw except Image
+
+export pixie
+
+export staticglfw except Image
 
 var
   dpi: float32 = 1.0