--- a/src/pixie/demo.nim
+++ b/src/pixie/demo.nim
@@ -1,8 +1,8 @@
 import opengl, pixie, pixie/context
 
-import staticglfw except Image
-
 export pixie
+
+import staticglfw except Image
 
 export staticglfw except Image