--- a/tests/fuzz_leaks2.nim
+++ b/tests/fuzz_leaks2.nim
@@ -1,4 +1,4 @@
-import pixie, random, os
+import os, pixie, random
 
 when not defined(pixieLeakCheck):
   quit("Requires -d:pixieLeakCheck")