--- a/tests/fuzz_leaks3.nim
+++ b/tests/fuzz_leaks3.nim
@@ -1,4 +1,4 @@
-import pixie, random, pixie/fileformats/svg
+import pixie, pixie/fileformats/svg, random
 
 when not defined(pixieLeakCheck):
   quit("Requires -d:pixieLeakCheck")
@@ -8,7 +8,7 @@ randomize()
 let data = readFile("tests/images/svg/Ghostscript_Tiger.svg")
 
 for i in 0 ..< 100_000:
-  var image = decodeSvg(data, rand(300 .. 1800),  rand(30 .. 1800))
+  var image = decodeSvg(data, rand(300 .. 1800), rand(30 .. 1800))
 
   # image.writeFile("tests/fuzz_leaks3.png")
   # break