… examples/blur.png
… examples/gradient.png
… examples/heart.png
… examples/image_tiled.png
… examples/line.png
… examples/masking.png
… examples/rounded_rectangle.png
… examples/shadow.png
… examples/square.png
… examples/text.png
… examples/text_spans.png
… examples/tiger.png
… tests/benchmark_text.nim
… tests/fonts/diffs/spans6.png
… tests/fonts/diffs/strikethrough1.png
… tests/fonts/diffs/strikethrough2.png
… tests/fonts/diffs/strikethrough3.png
… tests/fonts/diffs/underline1.png
… tests/fonts/diffs/underline2.png
… tests/fonts/diffs/underline3.png
… tests/fonts/image_stroke.png
… tests/fonts/mask_stroke.png
… tests/fonts/masters/spans6.png
… tests/fonts/masters/strikethrough1.png
… tests/fonts/masters/strikethrough2.png
… tests/fonts/masters/strikethrough3.png
… tests/fonts/masters/underline1.png
… tests/fonts/masters/underline2.png
… tests/fonts/masters/underline3.png
… tests/fonts/rendered/spans6.png