before:
`svg decode ........................ 19.549 ms     19.692 ms    ±0.135   x254`

after:
`svg decode ........................ 18.989 ms     19.147 ms    ±0.146   x261`

slight image changes due to sort difference imo (when ==, which one goes first changed sometimes results in slight blend change of who is source vs backdrop)